gate download of web-studio-sdk for most visitors who don't need it
remove bundle size of first load
