+++ components/agility-components/GetADemo/GetADemo.client.tsx
@@ -1,7 +1,7 @@
 "use client"
 
 import Script from "next/script"
-import { posthog } from "posthog-js"
+import { capture, identify } from "lib/analytics/posthog"
 import { useCallback, useEffect, useRef, useState } from "react"
 import { useRouter } from "next/navigation"
 
@@ -50,10 +50,10 @@ export const GetADemoClient = ({ hubspotForm, redirectURL }: Props) => {
 				console.log("Form submitted successfully:", name)
 				const emailAddress = data?.submissionValues?.email
 				if (emailAddress) {
-					posthog.identify(emailAddress)
+					identify(emailAddress)
 				}
 
-				posthog.capture("website-form-submission", {
+				capture("website-form-submission", {
 					name: name
 				})
