+++ components/search/SearchDialog.tsx
@@ -1,0 +1,319 @@
+'use client'
+
+import { liteClient } from 'algoliasearch/lite'
+import { getAlgoliaResults } from '@algolia/autocomplete-preset-algolia'
+import {
+	useCallback,
+	useEffect,
+	useId,
+	useRef,
+	useState,
+} from 'react'
+import { usePathname, useRouter, useSearchParams } from 'next/navigation'
+import {
+	type AutocompleteState,
+	createAutocomplete,
+} from '@algolia/autocomplete-core'
+import { Dialog, DialogPanel, DialogBackdrop } from '@headlessui/react'
+import clsx from 'clsx'
+import SearchResults from './SearchResults'
+import { SearchInput } from './SearchInput'
+import type { Autocomplete, Result } from './Search'
+
+type EmptyObject = Record<string, never>
+
+const searchClient = liteClient(
+	process.env.NEXT_PUBLIC_ALGOLIA_APP_ID || '',
+	process.env.NEXT_PUBLIC_ALGOLIA_SEARCH_API_KEY || ''
+)
+
+function useAutocomplete({ close }: { close: () => void }) {
+	let id = useId()
+	let router = useRouter()
+	let [autocompleteState, setAutocompleteState] = useState<
+		AutocompleteState<Result> | EmptyObject
+	>({})
+
+	function navigate({ item }: { item: Result }) {
+		const itemUrl = item.url
+		if (!itemUrl) {
+			return
+		}
+
+		let url = itemUrl
+		if (url.startsWith('https://agilitycms.com')) {
+			url = url.replace('https://agilitycms.com', '')
+		} else if (url.startsWith('/')) {
+			url = `https://agilitycms.com/docs${url}`
+		}
+
+		router.push(url)
+
+		if (
+			itemUrl ===
+			window.location.pathname + window.location.search + window.location.hash
+		) {
+			close()
+		}
+	}
+
+	let [autocomplete] = useState<Autocomplete>(() =>
+		createAutocomplete<
+			Result,
+			React.SyntheticEvent,
+			React.MouseEvent,
+			React.KeyboardEvent
+		>({
+			id,
+			placeholder: 'Search...',
+			defaultActiveItemId: 0,
+			onStateChange({ state }) {
+				setAutocompleteState(state)
+			},
+			shouldPanelOpen({ state }) {
+				return state.query !== ''
+			},
+			navigator: {
+				navigate,
+			},
+			insights: true,
+			getSources({ query }) {
+				return [
+					{
+						sourceId: 'documentation',
+						getItems() {
+							return getAlgoliaResults({
+								searchClient,
+								queries: [
+									{
+										indexName: 'agility-website',
+										params: {
+											query,
+											hitsPerPage: 10,
+										},
+									},
+									{
+										indexName: 'doc_site',
+										params: {
+											query,
+											hitsPerPage: 10,
+										},
+									},
+								],
+							})
+						},
+						getItemUrl({ item }) {
+							return item.url
+						},
+						onSelect: navigate,
+					},
+				]
+			},
+		})
+	)
+
+	return { autocomplete, autocompleteState }
+}
+
+export function SearchDialog({
+	open,
+	setOpen,
+	className,
+}: {
+	open: boolean
+	setOpen: (open: boolean) => void
+	className?: string
+}) {
+	let formRef = useRef<React.ElementRef<'form'>>(null)
+	let panelRef = useRef<React.ElementRef<'div'>>(null)
+	let inputRef = useRef<React.ElementRef<typeof SearchInput>>(null)
+	let { autocomplete, autocompleteState } = useAutocomplete({
+		close() {
+			setOpen(false)
+		},
+	})
+	let pathname = usePathname()
+	let searchParams = useSearchParams()
+	let router = useRouter()
+
+	const [extraResults, setExtraResults] = useState<Result[]>([])
+	const [isLoadingMore, setIsLoadingMore] = useState(false)
+	const [totalHits, setTotalHits] = useState<number | null>(null)
+	const pageRef = useRef(0)
+	const queryRef = useRef('')
+	const isLoadingMoreRef = useRef(false)
+	const hasMoreRef = useRef(true)
+
+	useEffect(() => {
+		setOpen(false)
+	}, [pathname, searchParams, setOpen])
+
+	useEffect(() => {
+		const query = (autocompleteState as AutocompleteState<Result>).query || ''
+		if (query !== queryRef.current) {
+			queryRef.current = query
+			pageRef.current = 0
+			isLoadingMoreRef.current = false
+			hasMoreRef.current = true
+			setExtraResults([])
+			setIsLoadingMore(false)
+			setTotalHits(null)
+
+			if (panelRef.current) {
+				panelRef.current.scrollTop = 0
+			}
+
+			if (query) {
+				searchClient.search({
+					requests: [
+						{ indexName: 'agility-website', query, hitsPerPage: 0 },
+						{ indexName: 'doc_site', query, hitsPerPage: 0 },
+					]
+				}).then((response: any) => {
+					if (queryRef.current === query) {
+						const total = response.results.reduce(
+							(sum: number, r: any) => sum + (r.nbHits || 0), 0
+						)
+						setTotalHits(total)
+					}
+				}).catch(() => { })
+			}
+		}
+	}, [autocompleteState])
+
+	function navigateToResult(result: Result) {
+		let url = result.url
+		if (!url) return
+		if (url.startsWith('https://agilitycms.com')) {
+			url = url.replace('https://agilitycms.com', '')
+		} else if (url.startsWith('/')) {
+			url = `https://agilitycms.com/docs${url}`
+		}
+		router.push(url)
+		setOpen(false)
+	}
+
+	const loadMore = useCallback(async () => {
+		const query = queryRef.current
+		if (!query || isLoadingMoreRef.current || !hasMoreRef.current) return
+
+		isLoadingMoreRef.current = true
+		setIsLoadingMore(true)
+		const nextPage = pageRef.current + 1
+
+		try {
+			const response: any = await searchClient.search({
+				requests: [
+					{ indexName: 'agility-website', query, hitsPerPage: 10, page: nextPage },
+					{ indexName: 'doc_site', query, hitsPerPage: 10, page: nextPage },
+				]
+			})
+
+			if (queryRef.current !== query) return
+
+			const newResults: Result[] = []
+			for (let i = 0; i < response.results.length; i++) {
+				const indexName = i === 0 ? 'agility-website' : 'doc_site'
+				for (const hit of (response.results[i].hits || [])) {
+					newResults.push({
+						url: hit.url,
+						title: hit.title,
+						description: hit.description,
+						ogImage: hit.ogImage,
+						section: hit.section,
+						concept: hit.concept,
+						publishedDate: hit.publishedDate,
+						contentType: hit.contentType,
+						__autocomplete_indexName: indexName,
+					})
+				}
+			}
+
+			if (newResults.length === 0) {
+				hasMoreRef.current = false
+			} else {
+				const idxA = newResults.filter(r => r.__autocomplete_indexName !== 'doc_site')
+				const idxB = newResults.filter(r => r.__autocomplete_indexName === 'doc_site')
+				const interleaved: Result[] = []
+				const maxLen = Math.max(idxA.length, idxB.length)
+				for (let i = 0; i < maxLen; i++) {
+					if (i < idxA.length) interleaved.push(idxA[i])
+					if (i < idxB.length) interleaved.push(idxB[i])
+				}
+				setExtraResults(prev => [...prev, ...interleaved])
+				pageRef.current = nextPage
+			}
+		} catch {
+			// ignore errors
+		} finally {
+			if (queryRef.current === query) {
+				isLoadingMoreRef.current = false
+				setIsLoadingMore(false)
+			}
+		}
+	}, [])
+
+	const handlePanelScroll = useCallback((e: React.UIEvent<HTMLDivElement>) => {
+		const { scrollTop, scrollHeight, clientHeight } = e.currentTarget
+		if (scrollHeight - scrollTop - clientHeight < 150) {
+			loadMore()
+		}
+	}, [loadMore])
+
+	return (
+		<Dialog
+			open={open}
+			onClose={() => {
+				setOpen(false)
+				autocomplete.setQuery('')
+			}}
+			className={clsx('fixed inset-0 z-50', className)}
+		>
+			<DialogBackdrop
+				transition
+				className="fixed inset-0 bg-zinc-400/25 backdrop-blur-sm data-[closed]:opacity-0 data-[enter]:duration-300 data-[leave]:duration-200 data-[enter]:ease-out data-[leave]:ease-in "
+			/>
+
+			<div className="fixed inset-0 overflow-y-auto px-4 py-4 sm:px-6 sm:py-20 md:py-32 lg:px-8 lg:py-[15vh]">
+				<DialogPanel
+					transition
+					className="mx-auto transform-gpu overflow-hidden rounded-lg bg-zinc-50 shadow-xl ring-1 ring-zinc-900/7.5 data-[closed]:scale-95 data-[closed]:opacity-0 data-[enter]:duration-300 data-[leave]:duration-200 data-[enter]:ease-out data-[leave]:ease-in sm:max-w-xl  "
+				>
+					<div {...autocomplete.getRootProps({})}>
+						<form
+							ref={formRef}
+							{...autocomplete.getFormProps({
+								inputElement: inputRef.current,
+							})}
+						>
+							<SearchInput
+								ref={inputRef}
+								autocomplete={autocomplete}
+								autocompleteState={autocompleteState}
+								onClose={() => setOpen(false)}
+							/>
+							<div
+								ref={panelRef}
+								className="border-t border-zinc-200 bg-white empty:hidden max-h-[70vh] overflow-y-auto"
+								{...autocomplete.getPanelProps({})}
+								onScroll={handlePanelScroll}
+							>
+								{autocompleteState.isOpen && (
+									<SearchResults
+										autocomplete={autocomplete}
+										query={autocompleteState.query}
+										collection={autocompleteState.collections[0]}
+										extraResults={extraResults}
+										isLoadingMore={isLoadingMore}
+										totalHits={totalHits}
+										onExtraResultClick={navigateToResult}
+									/>
+								)}
+							</div>
+						</form>
+					</div>
+				</DialogPanel>
+			</div>
+		</Dialog>
+	)
+}
